refactor: Car 정보 폴더 생성 후 클래스 이름 변경
