+++ src/main/java/racingcar/Application.java
@@ -1,8 +1,10 @@
 package racingcar;
 
+import racingcar.car.UserInputCarName;
+
 public class Application {
     public static void main(String[] args) {
-        User user = new User();
+        UserInputCarName user = new UserInputCarName();
         user.splitByComma(user.inputName());
     }
 }
